requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
python-docx==1.1.0
aiosqlite==0.19.0
deep-translator==1.11.4
aiohttp==3.9.1

